// generated/src/app/api-models/service-area-post-cd.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ServiceAreaPostCd {
  serviceAreaName: string;
  seqPostCode: number;
  fromPostCode: string;
  thruPostCode: string;
  operator: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}